+++ sources/magic.cpp
@@ -3,7 +3,6 @@
 
 #include <cmath>
 
-#include <algorithm>
 #include <array>
 #include <utility>